Split parsing into own file
